Merge from upstream

+++ src/kudu/tserver/ts_tablet_manager.cc
@@ -559,8 +559,9 @@ Status TSTabletManager::DeleteTablet(
   return Status::OK();
 }
 
-string TSTabletManager::LogPrefix(const string& tablet_id) const {
-  return "T " + tablet_id + " P " + fs_manager_->uuid() + ": ";
+string TSTabletManager::LogPrefix(const string& tablet_id, FsManager *fs_manager) {
+  DCHECK(fs_manager != nullptr);
+  return Substitute("T $0 P $1: ", tablet_id, fs_manager->uuid());
 }
 
 Status TSTabletManager::CheckRunningUnlocked(
@@ -887,17 +888,24 @@ Status TSTabletManager::HandleNonReadyTabletOnStartup(const scoped_refptr<Tablet
       << "Unexpected TabletDataState in tablet " << tablet_id << ": "
       << TabletDataState_Name(data_state) << " (" << data_state << ")";
 
-  // Roll forward deletions, as needed.
-  LOG(WARNING) << LogPrefix(tablet_id) << "Tablet Manager startup: Rolling forward tablet deletion "
-               << "of type " << TabletDataState_Name(data_state);
+  // If the tablet is already fully tombstoned with no remaining data or WAL,
+  // then no need to roll anything forward.
+  bool skip_deletion = meta->IsTombstonedWithNoBlocks() &&
+      !Log::HasOnDiskData(meta->fs_manager(), tablet_id);
+
+  LOG_IF(WARNING, !skip_deletion)
+      << LogPrefix(tablet_id) << "Tablet Manager startup: Rolling forward tablet deletion "
+      << "of type " << TabletDataState_Name(data_state);
 
   if (data_state == TABLET_DATA_COPYING) {
     // We tombstone tablets that failed to copy.
     data_state = TABLET_DATA_TOMBSTONED;
   }
 
-  // Passing no OpId will retain the last_logged_opid that was previously in the metadata.
-  RETURN_NOT_OK(DeleteTabletData(meta, data_state, boost::none));
+  if (!skip_deletion) {
+    // Passing no OpId will retain the last_logged_opid that was previously in the metadata.
+    RETURN_NOT_OK(DeleteTabletData(meta, data_state, boost::none));
+  }
 
   // Register TOMBSTONED tablets so that they get reported to the Master, which
   // allows us to permanently delete replica tombstones when a table gets
@@ -913,7 +921,8 @@ Status TSTabletManager::DeleteTabletData(const scoped_refptr<TabletMetadata>& me
                                          TabletDataState data_state,
                                          const boost::optional<OpId>& last_logged_opid) {
   const string& tablet_id = meta->tablet_id();
-  LOG(INFO) << LogPrefix(tablet_id) << "Deleting tablet data with delete state "
+  LOG(INFO) << LogPrefix(tablet_id, meta->fs_manager())
+            << "Deleting tablet data with delete state "
             << TabletDataState_Name(data_state);
   CHECK(data_state == TABLET_DATA_DELETED ||
         data_state == TABLET_DATA_TOMBSTONED)
@@ -923,7 +932,8 @@ Status TSTabletManager::DeleteTabletData(const scoped_refptr<TabletMetadata>& me
   // Note: Passing an unset 'last_logged_opid' will retain the last_logged_opid
   // that was previously in the metadata.
   RETURN_NOT_OK(meta->DeleteTabletData(data_state, last_logged_opid));
-  LOG(INFO) << LogPrefix(tablet_id) << "Tablet deleted. Last logged OpId: "
+  LOG(INFO) << LogPrefix(tablet_id, meta->fs_manager())
+            << "Tablet deleted. Last logged OpId: "
             << meta->tombstone_last_logged_opid();
   MAYBE_FAULT(FLAGS_fault_crash_after_blocks_deleted);
 
@@ -946,14 +956,13 @@ void TSTabletManager::LogAndTombstone(const scoped_refptr<TabletPeer>& peer,
                                       const std::string& msg,
                                       const Status& s) {
   const string& tablet_id = peer->tablet_id();
-  const string kLogPrefix = "T " + tablet_id + " P " + fs_manager_->uuid() + ": ";
-  LOG(WARNING) << kLogPrefix << msg << ": " << s.ToString();
+  LOG(WARNING) << LogPrefix(tablet_id) << msg << ": " << s.ToString();
 
   Status delete_status = DeleteTabletData(
       peer->tablet_metadata(), TABLET_DATA_TOMBSTONED, boost::optional<OpId>());
   if (PREDICT_FALSE(!delete_status.ok())) {
     // This failure should only either indicate a bug or an IO error.
-    LOG(FATAL) << kLogPrefix << "Failed to tombstone tablet after tablet copy: "
+    LOG(FATAL) << LogPrefix(tablet_id) << "Failed to tombstone tablet after tablet copy: "
                << delete_status.ToString();
   }
   peer->StatusMessage(Substitute("Tombstoned tablet: $0 ($1)", msg, s.ToString()));
